requirements.txt
# Flask and web framework dependencies
flask==2.3.3
werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2

# Database connectivity
pymysql==1.1.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.25

# Authentication and security
flask-bcrypt==1.0.1
bcrypt==4.0.1

# Environment and configuration
python-dotenv==1.0.0

# File handling
werkzeug==2.3.7

# Utilities
uuid==1.30

# Flask extensions
flask-mail==0.9.1

# Task processing and caching
celery==5.3.6
redis==5.0.1

# AI integration
openai==1.5.0